requirements.txt
streamlit>=1.32
google-generativeai>=0.8.0
protobuf>=4.25.3,<5
python-dotenv
